WIP: add remove item functionality to cart
